3. 패키지 이름 변경: global -> common
