+++ src/main/java/com/nexters/phochak/auth/application/AppleOAuthServiceImpl.java
@@ -3,12 +3,11 @@ package com.nexters.phochak.auth.application;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nexters.phochak.auth.presentation.AppleAuthKeyFeignClient;
+import com.nexters.phochak.common.exception.PhochakException;
+import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.dto.AppleUserInformation;
 import com.nexters.phochak.dto.OAuthUserInformation;
 import com.nexters.phochak.dto.jwt.Keys;
-import com.nexters.phochak.exception.PhochakException;
-import com.nexters.phochak.exception.ResCode;
-import com.nexters.phochak.auth.application.OAuthService;
 import com.nexters.phochak.specification.OAuthProviderEnum;
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.crypto.RSASSAVerifier;
