+++ src/main/java/com/nexters/phochak/auth/application/JwtTokenServiceImpl.java
@@ -1,12 +1,12 @@
 package com.nexters.phochak.auth.application;
 
 import com.auth0.jwt.JWT;
-import com.nexters.phochak.global.config.property.JwtProperties;
+import com.nexters.phochak.common.config.property.JwtProperties;
+import com.nexters.phochak.common.exception.PhochakException;
+import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.dto.TokenDto;
 import com.nexters.phochak.dto.request.ReissueTokenRequestDto;
 import com.nexters.phochak.dto.response.JwtResponseDto;
-import com.nexters.phochak.exception.PhochakException;
-import com.nexters.phochak.exception.ResCode;
 import com.nexters.phochak.repository.RefreshTokenRepository;
 import com.nexters.phochak.service.JwtTokenService;
 import io.jsonwebtoken.Claims;
