+++ src/main/java/com/nexters/phochak/common/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.global.config;
+package com.nexters.phochak.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
