+++ src/main/java/com/nexters/phochak/common/config/property/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.global.config.property;
+package com.nexters.phochak.common.config.property;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,9 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 @Getter
 @RequiredArgsConstructor
 @ConstructorBinding
-@ConfigurationProperties(prefix = "slack.report")
-public class SlackReportProperties {
-    private final String botNickname;
-    private final String webHookUri;
+@ConfigurationProperties(prefix = "jwt.token")
+public class JwtProperties {
+    private final String secretKey;
+    private final long accessTokenExpireLength;
+    private final long refreshTokenExpireLength;
 }
