+++ src/main/java/com/nexters/phochak/common/config/property/KakaoLoginProperties.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.global.config.property;
+package com.nexters.phochak.common.config.property;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,9 +8,8 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 @Getter
 @RequiredArgsConstructor
 @ConstructorBinding
-@ConfigurationProperties(prefix = "jwt.token")
-public class JwtProperties {
-    private final String secretKey;
-    private final long accessTokenExpireLength;
-    private final long refreshTokenExpireLength;
+@ConfigurationProperties(prefix = "kakao")
+public class KakaoLoginProperties {
+    private final String clientId;
+    private final String redirectUri;
 }
