+++ src/main/java/com/nexters/phochak/common/config/property/RedisProperties.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.global.config.property;
+package com.nexters.phochak.common.config.property;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,9 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 @Getter
 @RequiredArgsConstructor
 @ConstructorBinding
-@ConfigurationProperties(prefix = "kakao")
-public class KakaoLoginProperties {
-    private final String clientId;
-    private final String redirectUri;
+@ConfigurationProperties(prefix = "spring.redis")
+public class RedisProperties {
+    private final String host;
+    private final int port;
+    private final String password;
 }
