+++ src/main/java/com/nexters/phochak/common/exception/ResCode.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.exception;
+package com.nexters.phochak.common.exception;
 
 public enum ResCode {
