+++ src/main/java/com/nexters/phochak/domain/Post.java
@@ -1,6 +1,6 @@
 package com.nexters.phochak.domain;
 
-import com.nexters.phochak.global.domain.BaseTime;
+import com.nexters.phochak.common.domain.BaseTime;
 import com.nexters.phochak.specification.PostCategoryEnum;
 import lombok.Builder;
 import lombok.Getter;
