+++ src/main/java/com/nexters/phochak/domain/User.java
@@ -1,6 +1,6 @@
 package com.nexters.phochak.domain;
 
-import com.nexters.phochak.global.domain.BaseTime;
+import com.nexters.phochak.common.domain.BaseTime;
 import com.nexters.phochak.specification.OAuthProviderEnum;
 import lombok.Builder;
 import lombok.Getter;
