+++ src/main/java/com/nexters/phochak/service/impl/PostServiceImpl.java
@@ -2,6 +2,8 @@ package com.nexters.phochak.service.impl;
 
 import com.nexters.phochak.auth.UserContext;
 import com.nexters.phochak.client.StorageBucketClient;
+import com.nexters.phochak.common.exception.PhochakException;
+import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.domain.Post;
 import com.nexters.phochak.domain.User;
 import com.nexters.phochak.dto.HashtagFetchDto;
@@ -11,11 +13,8 @@ import com.nexters.phochak.dto.PostFetchDto;
 import com.nexters.phochak.dto.PostUploadKeyResponseDto;
 import com.nexters.phochak.dto.request.CustomCursor;
 import com.nexters.phochak.dto.request.PostCreateRequestDto;
-import com.nexters.phochak.dto.request.PostFilter;
 import com.nexters.phochak.dto.request.PostUpdateRequestDto;
 import com.nexters.phochak.dto.response.PostPageResponseDto;
-import com.nexters.phochak.exception.PhochakException;
-import com.nexters.phochak.exception.ResCode;
 import com.nexters.phochak.repository.HashtagRepository;
 import com.nexters.phochak.repository.PostRepository;
 import com.nexters.phochak.repository.ShortsRepository;
