+++ src/main/java/com/nexters/phochak/service/impl/ReportPostServiceImpl.java
@@ -1,10 +1,10 @@
 package com.nexters.phochak.service.impl;
 
+import com.nexters.phochak.common.exception.PhochakException;
+import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.domain.Post;
 import com.nexters.phochak.domain.ReportPost;
 import com.nexters.phochak.domain.User;
-import com.nexters.phochak.exception.PhochakException;
-import com.nexters.phochak.exception.ResCode;
 import com.nexters.phochak.repository.PostRepository;
 import com.nexters.phochak.repository.ReportPostRepository;
 import com.nexters.phochak.repository.UserRepository;
