+++ src/test/java/com/nexters/phochak/service/NCPShortsServiceTest.java
@@ -1,6 +1,6 @@
 package com.nexters.phochak.service;
 
-import com.nexters.phochak.global.config.property.NCPStorageProperties;
+import com.nexters.phochak.common.config.property.NCPStorageProperties;
 import com.nexters.phochak.domain.Post;
 import com.nexters.phochak.domain.Shorts;
 import com.nexters.phochak.domain.User;
